Move scraper reporting to HTTP and extract register module
